Fixed player shoot delay by creating ShootDelayTimer class.

+++ src/Model/SpriteAnimation.java
@@ -11,18 +11,18 @@ public class SpriteAnimation extends Transition {
     private final ImageView imageView;
     private final int count;
     private final int columns;
-    private final double offsetX;
-    private final double offsetY;
+    private final int offsetX;
+    private final int offsetY;
     private final int width;
     private final int height;
 
-    private double lastIndex;
+    private int lastIndex;
 
     public SpriteAnimation(
             ImageView imageView,
             Duration duration,
             int count,   int columns,
-            double offsetX, double offsetY,
+            int offsetX, int offsetY,
             int width,   int height) {
         this.imageView = imageView;
         this.count     = count;
@@ -36,10 +36,10 @@ public class SpriteAnimation extends Transition {
     }
 
     protected void interpolate(double k) {
-        final double index = Math.min((int) Math.floor(k * count), count - 1);
+        final int index = Math.min((int) Math.floor(k * count), count - 1);
         if (index != lastIndex) {
-            final double x = (index % columns) * width  + offsetX;
-            final double y = (index / columns) * height + offsetY;
+            final int x = (index % columns) * width  + offsetX;
+            final int y = (index / columns) * height + offsetY;
             imageView.setViewport(new Rectangle2D(x, y, width, height));
             lastIndex = index;
         }
